Fixed deprecation in ScreenshotTestRule.| #2560
